Moved word of the day to a separate view
